requirements.txt
aiogram==3.7.0
aiohttp
apscheduler
python-dotenv
openai>=1.0.0
pytz
yookassa
